requirements.txt
anyio>=4.8.0,<5.0.0
beautifulsoup4~=4.13.3
certifi==2024.8.30
charset-normalizer==3.2.0
EasyProcess==1.1
entrypoint2==1.1
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.8
MouseInfo==0.1.3
mss==9.0.1
Pillow==10.4.0
pyaes==1.6.1
pyasn1==0.5.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
Pyrogram==2.0.106
pyscreenshot==3.1
PyScreeze==0.1.29
PySocks==1.7.1
pyTelegramBotAPI==4.12.0
python-telegram-bot==20.4
pytweening==1.0.7
requests~=2.32.3
rsa==4.9
rubicon-objc==0.4.6
sniffio==1.3.0
soupsieve==2.5
telebot==0.0.5
Telethon==1.28.5
urllib3==2.2.2
schedule~=1.2.0
annotated-types==0.6.0
colorama==0.4.6
distro==1.8.0
google-generativeai>=0.3.0
pydantic==2.9.2
pydantic-core==2.23.4
tqdm==4.66.1
cryptography==43.0.1
pycparser==2.21
openpyxl==3.1.2
spotdl~=4.2.11
APScheduler~=3.10.4
spotipy~=2.24.0
google~=3.0.0
protobuf>=3.19.5,<5.0.0
python-dotenv~=1.0.0
tenacity~=8.2.3
psycopg2~=2.9.0

# Development dependencies for improved code quality
mypy==1.5.1
flake8==6.0.0
pytest==7.4.0
pytest-mock==3.11.1
